removes RHW screwy fixer and rewrites DxD/OxD INRULs
The INRUL code has been significantly simplified and is also more
stable, as it dispenses with the screwy RUL2 fixer code. This makes
overrides of these intersections more stable.

The same still needs to be applied to Road and Onewayroad. I have the
code ready, including diagonal T intersections, but as there are
currently no override networks for those, there's no hurry.