Extra RUL2 for flextrack
This update contains RUL2 for the fixup of the override of the B2
diagonal Override to B1 STR diagonal from the end tile. This will not be
the only update unfortunately. I encourage everyone who knows how to
compile an installer and run the requisite files to do so and report
back. -eggman121