Add prevent for Road/OWR DxO half-dragged +-int
With the change to RUL1 that calls the T-int texture 0x09703a00 for the half-drag or T-int version of a DxO, this change prevents the user from forming full +-ints using the half drag method.

The reason for this is to avoid having +-int formed using the wrong textures.  The half-dragging method in this instance does not have any obvious benefit.  This change ensures that the +-int will always have the correct textures (0x09703900) at the expense of disabling half-drag construction.