Tile Drop FLEX-WT 8S/6S fix
Some lines in the "8S-6S -> 10S-8S Stability" section of the Sec7h8_FLEXWT files were mixed up between LHD and RHD.  The error was in the LHD code but was affecting the RHD setups.  After determining the correct code for RHD vs LHD, I discovered the lines were duplicate of previous lines in the "7 to Orth" section of "L0 RHW-8S/L0 RHW-6S Transition" in the same file.

It is not clear if the lines were intended to do something specific for "8S-6S -> 10S-8S Stability", but deleting them solves this problem and doesn't seem to introduce new issues.

 Progress on #345