drop fuzzy translations in conversion to dat file
Translations marked as fuzzy are really void, as they need to be
reconfirmed or retranslated manually. Therefore, they must be ignored
when creating the dat file, as they are potentially completely wrong.

The implementation involves a workaround for a bug in the po parser
library, which is added to the test cases.

This commit also improves the error reporting for duplicate TGIs in .po
files.