Add stability to L1 RHW-12S D1 Ramp FHT adjacency
Necessary because the base ramp (RHW-3) emits override to the up transition from the top of the ramp.  This ramp needs to override that.

There is still an issue with the FHT code though.  Currently in this situation, the 12S override does not propagate through the FHT ramp - the 12S version of the ramp is not overriding the RHW-3 version of the ramp.

Possible solution to that issue would be adding more code to the FHT RUL2 to account for this, or to eliminate the RHW-3 override from the base D1 ramp which would eliminate the need for this commit.