generalize FLEX ramps for use next to OSTs
This (re-)adds support for FLEX ramps next to OSTs, for orthogonal
outside ramps of type A1, B1, D1, E1.

This masks the single-tile side of the ramp by `0,?,0,?` so that it
matches both `0,2,0,2` and `0,4,0,4`. The latter occurs with OSTs. (In
case the question marks cause problems, we may need to duplicate all the
ramps for compatibility with both 2- and 4-flags.)