refactoring: re-use `context.preprocess` mechanism for `withSharedDiagonals` implementation
This uses the existing remapping mechanism instead of adding another
one. This works by locally overwriting the context.

In particular, this improves error messages, as stack traces will point
to the location where a rule is defined (`Rules += ...`) rather than
where a shared-tile diagonal was created in the remapping function.

Additionally, this re-implements the remapping function
`transformSharedDiagonals` to optimize performance.