define mirror variants for recently added OWR × Avenue intersections
Tiles with mirror variants (except for TLA networks) currently need to
be explicitly listed in `MirrorVariants.scala`.
Otherwise, the mirror variants aren't properly accounted for when RUL2
code is generated. (Detailed explanation: The RuleTransducer does not
know to automatically project the involved tiles and hence uses the
unprojected tiles. Consequently, the IDs of the unprojected tiles don't
resolve and, as a result, the generated RUL2 code does not contain the
expected rules.)

After this commit, resolving and transducing succeeds:

    transduce(Owr3~WE & Avenue~(1,0,0,0) | (Onewayroad~>Owr3)~WE)

    0x5104A309,1,0,0x09004B00,1,0=0x5104A309,1,0,0x51040000,1,0
    0x5104A309,1,0,0x09004B00,3,0=0x5104A309,1,0,0x51040000,3,0
    0x7104A309,3,1,0x09004B00,3,0=0x7104A309,3,1,0x51040000,3,0
    0x7104A309,3,1,0x09004B00,1,0=0x7104A309,3,1,0x51040000,1,0