parse semantic DLL version in Advisor Lua script
This allows to not show the advisor message if only the patch level is
higher than expected. For example:

- expected version: 1.2.0
- installed version: 1.2.3

This allows to release patches for the DLL without necessarily having to publish a new Lua script at the same time.
(Regardless, the Lua script should be updated with the following proper NAM release to ensure DLL versions without the patch are rejected, from then on.)

(Also increments the expected DLL version to 1.2.3.)